Rename server to temporal
