+++ temporal/server_option.go
@@ -1,4 +1,4 @@
-package server
+package temporal
 
 import (
 	"log"
